refactor(platforms): move Feishu platform module to platforms/feishu
Implements Phase 2 of #194 - Consolidate Feishu platform code

Changes:
- Move channels/platforms/feishu/* to platforms/feishu/
- Update all import paths to use new location
- Remove temporary README.md and re-export from platforms/feishu/index.ts
- Update channels/platforms/index.ts to re-export from platforms module

Benefits:
- Platform code consolidated in platforms/ directory
- Cleaner module organization
- Consistent with project architecture

Files moved:
- feishu-adapter.ts, feishu-adapter.test.ts
- feishu-message-sender.ts, feishu-message-sender.test.ts
- feishu-file-handler.ts, feishu-file-handler.test.ts
- card-builders/

All 714 tests pass.

Closes #267
